Implement dynamic test addition support
The update allows for the dynamic addition of tests during suite execution. The respective tests can be added/removed dynamically via event handlers during the running of the suite. Additionally, adjustments were made to the Runner and Suite modules to facilitate this functionality, and a unit test was created to ensure the applied changes work as expected.